Add optional skip_issuer parameter to IdToken.verify!
This is especially useful when using Microsoft Entra ID common
endpoint, as the issuer could be from another tenant.

When using this parameter it is recommended to set the audience as this
stays the same even if the issuer is from another tenant.

Related https://github.com/omniauth/omniauth_openid_connect/issues/166

Closes #95