access_token: allow handling the /userinfo as a JWT token
Though user information is usually returned as plain JSON, the spec
does allow a JWT as well[1]:

> The UserInfo Claims MUST be returned as the members of a JSON object
unless a signed or encrypted response was requested during Client
Registration.

I'm working with a provider that only returns JWT and does not allow
Content-Type override. The workaround is to check the response's
content type and parse the JWT if relevant.

[1]: https://openid.net/specs/openid-connect-core-1_0.html#UserInfo